Updates README.md
The current Minimal Example does not allow you to connect to a wallet successfully and the error provided does not provide much details as to why. I had a hard time tracking down why I could not connect to a wallet by default for the example. This update informs users during setup that they need to do a little more work to get a wallet connected successfully before developing without errors.