Updates README.md (#103)
* Updates README.md

The current Minimal Example does not allow you to connect to a wallet successfully and the error provided does not provide much details as to why. I had a hard time tracking down why I could not connect to a wallet by default for the example. This update informs users during setup that they need to do a little more work to get a wallet connected successfully before developing without errors.

* Updating README.md

Updating links and content

* Removing 'proxyUrl's'

Co-authored-by: Alan Smithee <ggnore.alan.smithee@gmail.com>

* Removing proxy urls since blockfrost will be used by default in this example

Co-authored-by: Alan Smithee <ggnore.alan.smithee@gmail.com>

---------

Co-authored-by: Alan Smithee <ggnore.alan.smithee@gmail.com>